Revert "[IOTDB-5678] Introduce machine learning algorithm libraries on MLNode (#9338)"
This reverts commit 1bfcab297f90562258b8cc2bfe7929556c04cd7c.
