+++ mlnode/iotdb/mlnode/exception.py
@@ -16,7 +16,6 @@
 # under the License.
 #
 
-
 class _BaseError(Exception):
     """Base class for exceptions in this module."""
     pass
